+++ packages/mona-clients/mona-client-mini/src/index.ts
@@ -1,5 +1,5 @@
-export * from './apis';
 export * from './components';
 export { default as createApp } from './createApp';
 export { default as createPage } from './createPage';
 export { default as createMiniComponent } from './createMiniComponent';
+export { Mona } from './apis';
