+++ packages/mona-clients/mona-client-plugin/src/index.ts
@@ -1,5 +1,5 @@
-export * from './createPlugin';
 export * from './apis';
+export * from './createPlugin';
 export * from './components';
 export { createMiniComponent } from './createMiniComponent';
 export { createPluginLifeCycle } from './createPluginLifeCycle';
