+++ packages/mona-clients/mona-client-web/src/index.ts
@@ -1,8 +1,8 @@
-export * from './apis';
+export { Mona } from './apis';
 export * from './components';
 export * from './createWebApp';
 export { createMiniComponent } from './createMiniComponent';
-export { createAppLifeCycle, } from './createWebAppEvent';
+export { createAppLifeCycle } from './createWebAppEvent';
 export { createPageLifecycle } from './createWebPage';
 
 export { usePageEvent, useAppEvent, AppLifecycle, PageLifecycle } from '@bytedance/mona';
